break out path.ts from metamodel.ts; also move PathModel/PathOrEmptyModel to simple-models.ts (using an agent)
